paths: handle previously-unmounted hierarchies
If a user previously mounted a cgroup hierarchy on their local machine,
then unmounted it, the following conditions will be true:

1. This cgroup hierarchy may still have an entry in the
/proc/[pid]/cgroup file of processes on that machine
2. This cgroup hierarchy will have no mountpoint in
/proc/self/mountinfo.

These conditions would previously cause existingPath() to fail to find
any mountpoints for any cgroup hierarchies. This commit adds a check
for this condition, and causes existingPath() to ignore previously
unmounted hierarchies.